backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Task Queue
rq==1.15.1
redis==5.0.1

# AI & Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
pillow-simd==10.0.1.post0  # 效能優化版本

# Data Processing
numpy<2.0
pandas>=2.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing async endpoints

# Monitoring (Optional)
prometheus-client==0.19.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
